refactor: Make some minor adjustments - Update import paths - Pass env variables within scripts with consideration for all file systems
